Resolve "Create RAWG API game repository"

Closes #27

See merge request internationaal-project-06/server!21

+++ controllers/user_games_controller.js
@@ -1,11 +1,16 @@
 "use strict";
 
 const { AppError } = require('../errors');
-const discordRepository = require("../repositories/discord_repository");
 const userRepository = require("../repositories/user_repository");
 const userGamesRepository = require("../repositories/user_games_repository");
+const gameRepository = require("../repositories/game_repository");
 
-
+async function getGamesFromUser(userId) {
+    let gameIds = await userGamesRepository.getAllGamesFromUser(userId);
+    gameIds = gameIds.map(obj => obj.game_id);
+    let games = await Promise.all(gameIds.map(async id => gameRepository.getGame(id)));
+    return games;
+}
 
 async function addGameToBlackList(userId, gameId) {
     if(userId == null || gameId == null) {
@@ -85,6 +90,12 @@ async function resetBlacklist(userId) {
     return true;
 }
 
+async function hasGame(uid, gid) {
+    let userGames = await userGamesRepository.getAllGamesFromUser(uid);
+    userGames = userGames.map(game => game.game_id);
+    return userGames.includes(parseInt(gid));
+}
+
 async function connectGameToUser(uid, gid, hoursPlayed=0, rank=null) {
     if(uid == null || gid == null){
         throw new AppError(400, "Bad request");
@@ -92,6 +103,15 @@ async function connectGameToUser(uid, gid, hoursPlayed=0, rank=null) {
     if(await userRepository.getUserFromId(uid) == null) {
         throw new AppError(404, "User not found.");
     }
+    if(await gameRepository.getGame(gid) == null) {
+        // When the game doesn't exist in our database, try to add it
+        await gameRepository.addGame(gid);
+    }
+
+    if(await hasGame(uid, gid)) {
+        throw new AppError(400, "You already have this game.");
+    }
+
     return await userGamesRepository.connectGameToUser(uid, gid, hoursPlayed, rank);
 }
 
@@ -102,6 +122,11 @@ async function removeGameFromUser(uid, gid) {
     if(await userRepository.getUserFromId(uid) == null) {
         throw new AppError(404, "User not found.");
     }
+
+    if(!await hasGame(uid, gid)) {
+        throw new AppError(400, "You don't have this game");
+    }
+
     return await userGamesRepository.removeGameFromUser(uid, gid);
 
 }
@@ -110,6 +135,7 @@ module.exports = {
     addGameToBlackList,
     removeGameFromBlackList,
     resetBlacklist,
+    getGamesFromUser,
     connectGameToUser,
     removeGameFromUser,
     respondToMatchSuggestion
